Rename classes and packages
eu.agilejava.snoop -> eu.agilejava.snoopee
SnoopXX -> SnoopEEXX
