+++ snoopee-discovery/snoopee-client/src/main/java/eu/agilejava/snoopee/client/SnoopEEServiceUnavailableException.java
@@ -1,0 +1,54 @@
+/*
+ * The MIT License
+ *
+ * Copyright 2015 Ivar Grimstad (ivar.grimstad@gmail.com).
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package eu.agilejava.snoopee.client;
+
+import javax.ejb.ApplicationException;
+
+/**
+ * This exception is thrown if the snoop service is not available.
+ *
+ * @author Ivar Grimstad (ivar.grimstad@gmail.com)
+ */
+@ApplicationException
+public class SnoopEEServiceUnavailableException extends RuntimeException {
+
+    public SnoopEEServiceUnavailableException() {
+    }
+
+    public SnoopEEServiceUnavailableException(String message) {
+        super(message);
+    }
+
+    public SnoopEEServiceUnavailableException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public SnoopEEServiceUnavailableException(Throwable cause) {
+        super(cause);
+    }
+
+    public SnoopEEServiceUnavailableException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+}
