--- snoopee-discovery/snoopee-service/src/main/java/eu/agilejava/snoop/ui/SnoopController.java
@@ -1,56 +1,0 @@
-/*
- * The MIT License
- *
- * Copyright 2015 Ivar Grimstad (ivar.grimstad@gmail.com).
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package eu.agilejava.snoop.ui;
-
-import eu.agilejava.snoop.SnoopClientRegistry;
-import eu.agilejava.snoop.SnoopConfig;
-import java.util.Collection;
-import java.util.Set;
-import java.util.logging.Logger;
-import javax.ejb.EJB;
-import javax.enterprise.context.RequestScoped;
-import javax.inject.Named;
-
-/**
- * Controller for the snoop service UI.
- * 
- * @author Ivar Grimstad (ivar.grimstad@gmail.com)
- */
-@Named
-@RequestScoped
-public class SnoopController {
-
-   private static final Logger LOGGER = Logger.getLogger("eu.agilejava.snoop");
-
-   @EJB
-   private SnoopClientRegistry clients;
-
-   public Set<String> getClients() {
-      return clients.getClients();
-   }
-   
-   public Collection<SnoopConfig> getClientConfigurations() {
-      return clients.getServiceConfigs();
-   }
-}
