+++ snoopee-discovery/snoopee/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,1 +1,1 @@
-eu.agilejava.snoop.scan.SnoopScannerExtension
+eu.agilejava.snoopee.scan.SnoopEEScannerExtension
